tests/end2end: migrate tests to safe server code: UC55 and other
Relates to #925
