+++ tests/end2end/document/UC03_create_section/UC03_G1/UC03_G1_T01_section_with_empty_name/input/document.sdoc
