+++ tests/end2end/document/UC03_create_section/UC03_G1/UC03_G1_T02_malformed_rst_statement/input/document.sdoc
@@ -1,0 +1,6 @@
+[DOCUMENT]
+TITLE: Document 1
+
+[FREETEXT]
+Hello world!
+[/FREETEXT]
