+++ tests/end2end/document_tree/UC50_viewing_document_tree/UC50_T01_empty_document_tree/test_UC50_T01_empty_document_tree.py
@@ -10,23 +10,20 @@ path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
 
 class Test_UC50_T01_EmptyDocumentTree(BaseCase):
     def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
+        with SDocTestServer(
+            input_path=path_to_this_test_file_folder
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        test_server = SDocTestServer.create(path_to_sandbox)
-        test_server.run()
+            self.assert_text("PROJECT INDEX")
 
-        self.open(test_server.get_host_and_port())
+            self.assert_text("The document tree has no documents yet.")
 
-        self.assert_text("PROJECT INDEX")
-
-        self.assert_text("The document tree has no documents yet.")
-
-        self.assert_element(
-            (
-                "//*[@data-testid='document-tree-empty-text' "
-                "and contains(., 'The document tree has no documents yet.')]"
-            ),
-            by=By.XPATH,
-        )
+            self.assert_element(
+                (
+                    "//*[@data-testid='document-tree-empty-text' "
+                    "and "
+                    "contains(., 'The document tree has no documents yet.')]"
+                ),
+                by=By.XPATH,
+            )
