+++ tests/end2end/document_tree/UC51_creating_document/UC51_G1_validation/UC51_G1_T02_empty_document_path/test_UC51_G1_T02_empty_document_path.py
@@ -9,24 +9,20 @@ path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
 
 class Test_TC51_G1_T02_CreatingDocumentWithEmptyPath(BaseCase):
     def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
+        with SDocTestServer(
+            input_path=path_to_this_test_file_folder
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        test_server = SDocTestServer.create(path_to_sandbox)
-        test_server.run()
+            self.assert_text("PROJECT INDEX")
 
-        self.open(test_server.get_host_and_port())
+            self.assert_text("The document tree has no documents yet.")
 
-        self.assert_text("PROJECT INDEX")
+            self.click('[data-testid="tree-add-document-action"]')
 
-        self.assert_text("The document tree has no documents yet.")
+            self.type("#document_title", "Document 1")  # Empty document
+            self.type("#document_path", "")
 
-        self.click('[data-testid="tree-add-document-action"]')
+            self.click_xpath('//*[@data-testid="form-submit-action"]')
 
-        self.type("#document_title", "Document 1")  # Empty document
-        self.type("#document_path", "")
-
-        self.click_xpath('//*[@data-testid="form-submit-action"]')
-
-        self.assert_text("Document path must not be empty.")
+            self.assert_text("Document path must not be empty.")
