+++ tests/end2end/document_tree/UC51_creating_document/UC51_G1_validation/UC51_G1_T04_document_path_bad_chars/test_UC51_G1_T04_document_path_bad_chars.py
@@ -9,26 +9,22 @@ path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
 
 class Test_TC51_G1_T04_BadCharacters(BaseCase):
     def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
+        with SDocTestServer(
+            input_path=path_to_this_test_file_folder
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        test_server = SDocTestServer.create(path_to_sandbox)
-        test_server.run()
+            self.assert_text("PROJECT INDEX")
 
-        self.open(test_server.get_host_and_port())
+            self.assert_text("The document tree has no documents yet.")
 
-        self.assert_text("PROJECT INDEX")
+            self.click('[data-testid="tree-add-document-action"]')
 
-        self.assert_text("The document tree has no documents yet.")
+            self.type("#document_title", "Document 1")  # Empty document
+            self.type("#document_path", "docs/doc!#ument.sdoc")
 
-        self.click('[data-testid="tree-add-document-action"]')
-
-        self.type("#document_title", "Document 1")  # Empty document
-        self.type("#document_path", "docs/doc!#ument.sdoc")
-
-        self.click_xpath('//*[@data-testid="form-submit-action"]')
-        self.assert_text(
-            "Document path must be relative and only contain slashes, "
-            "alphanumeric characters, and underscore symbols."
-        )
+            self.click_xpath('//*[@data-testid="form-submit-action"]')
+            self.assert_text(
+                "Document path must be relative and only contain slashes, "
+                "alphanumeric characters, and underscore symbols."
+            )
