+++ tests/end2end/document_tree/UC51_creating_document/UC51_T01_creating_document/test_UC51_T01_creating_document.py
@@ -1,41 +1,26 @@
-import filecmp
-import os
-
 from seleniumbase import BaseCase
 
+from tests.end2end.end2end_test_setup import End2EndTestSetup
 from tests.end2end.server import SDocTestServer
-
-path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
 
 
 class Test_UC51_T01_CreatingDocument(BaseCase):
     def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
+        test_setup = End2EndTestSetup(path_to_test_file=__file__)
 
-        test_server = SDocTestServer.create(path_to_sandbox)
-        test_server.run()
+        with SDocTestServer(
+            input_path=test_setup.path_to_sandbox
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        self.open(test_server.get_host_and_port())
+            self.click_link("Add document")
+            self.click('[data-testid="tree-add-document-action"]')
 
-        self.click_link("Add document")
-        self.click('[data-testid="tree-add-document-action"]')
+            self.type("#document_title", "Document 1")
+            self.type("#document_path", "docs/document1.sdoc")
 
-        self.type("#document_title", "Document 1")
-        self.type("#document_path", "docs/document1.sdoc")
+            self.click_xpath('//*[@data-testid="form-submit-action"]')
 
-        self.click_xpath('//*[@data-testid="form-submit-action"]')
+            self.assert_text("Document 1")
 
-        self.assert_text("Document 1")
-
-        create_document_path = os.path.join(
-            path_to_sandbox, "docs", "document1.sdoc"
-        )
-        assert os.path.exists(create_document_path)
-        assert filecmp.cmp(
-            create_document_path,
-            os.path.join(
-                path_to_this_test_file_folder, "document.expected.sdoc"
-            ),
-        )
+        assert test_setup.compare_sandbox_and_expected_output()
