+++ tests/end2end/document_tree/UC55_import_document_from_reqif/UC55_T03_import_tree_of_two_documents/test_UC55_T03_import_tree_of_two_documents.py
@@ -1,8 +1,8 @@
-import filecmp
 import os
 
 from seleniumbase import BaseCase
 
+from tests.end2end.end2end_test_setup import End2EndTestSetup
 from tests.end2end.server import SDocTestServer
 
 path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
@@ -13,46 +13,24 @@ path_to_reqif_sample = os.path.join(
 
 class Test_UC55_T03_ImportTreeOfTwoDocuments(BaseCase):
     def test_01(self):
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
+        test_setup = End2EndTestSetup(path_to_test_file=__file__)
 
-        test_server = SDocTestServer.create(path_to_sandbox)
-        test_server.run()
+        with SDocTestServer(
+            input_path=test_setup.path_to_sandbox
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        self.open(test_server.get_host_and_port())
+            self.assert_text("PROJECT INDEX")
 
-        self.assert_text("PROJECT INDEX")
+            self.click('[data-testid="tree-import-reqif-action"]')
 
-        self.click('[data-testid="tree-import-reqif-action"]')
+            reqif_input_field = self.find_element(
+                "//*[@data-testid='form-reqif_file-field']"
+            )
+            reqif_input_field.send_keys(path_to_reqif_sample)
+            self.click_xpath('//*[@data-testid="form-submit-action"]')
 
-        reqif_input_field = self.find_element(
-            "//*[@data-testid='form-reqif_file-field']"
-        )
-        reqif_input_field.send_keys(path_to_reqif_sample)
-        self.click_xpath('//*[@data-testid="form-submit-action"]')
+            self.assert_text("Document 1")
+            self.assert_text("Document 2")
 
-        self.assert_text("Document 1")
-        self.assert_text("Document 2")
-
-        path_to_created_document1 = os.path.join(
-            path_to_sandbox, "Document_1.sdoc"
-        )
-        assert os.path.exists(path_to_created_document1)
-        assert filecmp.cmp(
-            path_to_created_document1,
-            os.path.join(
-                path_to_this_test_file_folder, "Document_1.expected.sdoc"
-            ),
-        )
-
-        path_to_created_document2 = os.path.join(
-            path_to_sandbox, "Document_2.sdoc"
-        )
-        assert os.path.exists(path_to_created_document2)
-        assert filecmp.cmp(
-            path_to_created_document2,
-            os.path.join(
-                path_to_this_test_file_folder, "Document_2.expected.sdoc"
-            ),
-        )
+        assert test_setup.compare_sandbox_and_expected_output()
