+++ tests/end2end/document_tree/UC56_export_tree_to_reqif/UC56_T01_export_tree_to_reqif/input/Document_1.sdoc
@@ -1,0 +1,15 @@
+[DOCUMENT]
+TITLE: Document 1
+
+[FREETEXT]
+Document 1 free text.
+[/FREETEXT]
+
+[SECTION]
+TITLE: Document 1 section
+
+[FREETEXT]
+Document 1 statement.
+[/FREETEXT]
+
+[/SECTION]
