+++ tests/end2end/document_tree/UC56_export_tree_to_reqif/UC56_T01_export_tree_to_reqif/input/Document_2.sdoc
@@ -1,0 +1,15 @@
+[DOCUMENT]
+TITLE: Document 2
+
+[FREETEXT]
+Document 2 free text.
+[/FREETEXT]
+
+[SECTION]
+TITLE: Document 2 section
+
+[FREETEXT]
+Document 2 statement.
+[/FREETEXT]
+
+[/SECTION]
