+++ tests/end2end/document_tree/UC56_export_tree_to_reqif/UC56_T01_export_tree_to_reqif/test_UC56_T01_export_tree_to_reqif.py
@@ -5,12 +5,9 @@ from sys import platform
 from seleniumbase import BaseCase
 
 from tests.end2end.conftest import DOWNLOADED_FILES_PATH
+from tests.end2end.end2end_test_setup import End2EndTestSetup
 from tests.end2end.server import DOWNLOAD_FILE_TIMEOUT, SDocTestServer
 
-path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
-path_to_expected_file = os.path.join(
-    path_to_this_test_file_folder, "export.expected.reqif"
-)
 path_to_expected_downloaded_file = os.path.join(
     DOWNLOADED_FILES_PATH, "export.reqif"
 )
@@ -20,31 +17,23 @@ class Test_UC56_T01_ExportTreeToReqIF(BaseCase):
     def test_01(self):
         shutil.rmtree(DOWNLOADED_FILES_PATH, ignore_errors=True)
 
-        path_to_sandbox = os.path.join(
-            path_to_this_test_file_folder, ".sandbox"
-        )
+        test_setup = End2EndTestSetup(path_to_test_file=__file__)
 
-        test_server = SDocTestServer.create(path_to_sandbox)
-        shutil.copyfile(
-            os.path.join(path_to_this_test_file_folder, "Document_1.sdoc"),
-            os.path.join(path_to_sandbox, "Document_1.sdoc"),
-        )
-        shutil.copyfile(
-            os.path.join(path_to_this_test_file_folder, "Document_2.sdoc"),
-            os.path.join(path_to_sandbox, "Document_2.sdoc"),
-        )
-        test_server.run()
+        with SDocTestServer(
+            input_path=test_setup.path_to_sandbox
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        self.open(test_server.get_host_and_port())
+            self.assert_text("Document 1")
+            self.assert_text("PROJECT INDEX")
 
-        self.assert_text("Document 1")
-        self.assert_text("PROJECT INDEX")
+            self.click('[data-testid="tree-export-reqif-action"]')
+            # FIXME: does not work on Linux CI
+            if platform == "linux" or platform == "linux2":
+                return
 
-        self.click('[data-testid="tree-export-reqif-action"]')
-        # FIXME: does not work on Linux CI
-        if platform == "linux" or platform == "linux2":
-            return
+            self.sleep(DOWNLOAD_FILE_TIMEOUT)
 
-        self.sleep(DOWNLOAD_FILE_TIMEOUT)
+            assert os.path.exists(path_to_expected_downloaded_file)
 
-        assert os.path.exists(path_to_expected_downloaded_file)
+        assert test_setup.compare_sandbox_and_expected_output()
