+++ tests/end2end/project_options/project_title/02_option_specified/test_02_option_specified.py
@@ -9,9 +9,9 @@ path_to_this_test_file_folder = os.path.dirname(os.path.abspath(__file__))
 
 class Test_ProjectOptions_ProjectTitle_01_OptionNotSpecified(BaseCase):
     def test_01(self):
-        test_server = SDocTestServer(input_path=path_to_this_test_file_folder)
-        test_server.run()
+        with SDocTestServer(
+            input_path=path_to_this_test_file_folder
+        ) as test_server:
+            self.open(test_server.get_host_and_port())
 
-        self.open(test_server.get_host_and_port())
-
-        self.assert_text("Test Project title read from TOML file")
+            self.assert_text("Test Project title read from TOML file")
